enhancement: refactor containers -> services directory
